src/ACME.PayrollManagement.Application/Helpers/FileHelper.cs
namespace ACME.PayrollManagement.Application.Helpers
{
    public static class FileHelper
    {
        public static string GetUniqueFileName(string fileName)
        {
            if (string.IsNullOrEmpty(fileName))
            {
                throw new ArgumentNullException(nameof(fileName));
            }

            fileName = Path.GetFileName(fileName);
            return string.Concat(Path.GetFileNameWithoutExtension(fileName)
                                , "_"
                                , Guid.NewGuid().ToString().AsSpan(0, 4)
                                , Path.GetExtension(fileName));
        }
    }
}